rename legacy project service to namespace
