+++ backend/lib/routes/index.js
@@ -17,7 +17,7 @@ module.exports = {
   '/cloudprofiles': require('./cloudprofiles'),
   '/seeds': require('./seeds'),
   '/gardenerextensions': require('./gardenerExtensions'),
-  '/namespaces': require('./namespaces'),
+  '/namespaces': require('./namespaces'), // TODO delete this legacy route
   '/namespaces/:namespace/shoots': require('./shoots'),
   '/namespaces/:namespace/tickets': require('./tickets'),
   '/namespaces/:namespace/cloudprovidersecrets': require('./cloudProviderSecrets'),
