+++ backend/lib/services/index.js
@@ -7,6 +7,7 @@
 module.exports = {
   cloudprofiles: require('./cloudprofiles'),
   seeds: require('./seeds'),
+  namespaces: require('./namespaces'),
   projects: require('./projects'),
   shoots: require('./shoots'),
   cloudProviderSecrets: require('./cloudProviderSecrets'),
